Initial check in of relatively stable Ark MySQL database schema (structure taken from ArkTest, data taken from ArkDev) on 22 Mar 2011.  This corresponds roughly to release 1.0.0 in JIRA's The-ARK project tracking (so the base may include some fixes/updates that would go into 1.0.1).
The folders are organised such that the schemas defined under "base" are supposed to correspond to a particular release ARK web-app version (in this case "20110322" was the name of today's version).  As such, an installation of a version of the web-app should use the corresponding base scripts to initialise their database.

"patch" will correspond to subsequent fixes/updates that will correspond to changes in the web-app towards the next release.

*NB: "base" scripts should never be used to upgrade the database as they involve dropping the relevant tables. For upgrades, the admin must apply all patches subsequent to the version of the "base" they used.

When the next release happens, we should create a new folder under "base" for the new version and re-extract the schema for this particular version (because such scripts are assumed to be used for fresh installations).

git-svn-id: svn://localhost@3081 35ee9b83-dd2c-47f4-99a2-c706670c48ce